+++ lib/sbmt/outbox/cli.rb
@@ -69,8 +69,8 @@ module Sbmt
         $stdout.puts AsciiArt.logo
         $stdout.puts "Outbox/Inbox worker has been started"
         $stdout.puts "Version: #{VERSION}"
-        $stdout.puts "Starting probes..."
         Sbmt::Outbox::Probes::Probe.run_probes
+        Sbmt::Outbox::Probes::Metrics.run_metrics
 
         worker.start
       end